requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
torch>=1.10.0
h5py>=3.6.0
PyYAML>=6.0

# Visualization
matplotlib>=3.5.0

# Training utilities
tqdm>=4.62.0
tensorboard>=2.8.0

# Optional: for Jupyter notebooks
# jupyter>=1.0.0
# ipywidgets>=7.6.0
